src/gojenga/requirements.txt
boto3==1.26.49
uvicorn==0.20.0
botocore==1.29.49
fastapi==0.89.1
pydantic==1.10.4
opentelemetry-distro==0.36b0
opentelemetry-exporter-jaeger-thrift==1.15.0
opentelemetry-instrumentation-fastapi==0.36b0
pytest==7.2.1
httpx==0.23.3
python-multipart==0.0.5